# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/assessment-templates/presets": {
      "get": {
        "operationId": "AssessmentTemplatesController_presets",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "assessment-templates"
        ]
      }
    },
    "/assessment-templates/presets/{key}/copy": {
      "post": {
        "operationId": "AssessmentTemplatesController_clonePreset",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "assessment-templates"
        ]
      }
    },
    "/assessment-templates": {
      "get": {
        "operationId": "AssessmentTemplatesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "assessment-templates"
        ]
      },
      "post": {
        "operationId": "AssessmentTemplatesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTemplateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "assessment-templates"
        ]
      }
    },
    "/assessment-templates/{id}": {
      "get": {
        "operationId": "AssessmentTemplatesController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "assessment-templates"
        ]
      },
      "patch": {
        "operationId": "AssessmentTemplatesController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTemplateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "assessment-templates"
        ]
      }
    },
    "/assessment-templates/{id}/publish": {
      "post": {
        "operationId": "AssessmentTemplatesController_publish",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "assessment-templates"
        ]
      }
    },
    "/assessment-templates/{id}/archive": {
      "post": {
        "operationId": "AssessmentTemplatesController_archive",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "assessment-templates"
        ]
      }
    },
    "/assessment-templates/{id}/restore": {
      "post": {
        "operationId": "AssessmentTemplatesController_restore",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "assessment-templates"
        ]
      }
    },
    "/assessments": {
      "get": {
        "operationId": "AssessmentsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "assessments"
        ]
      },
      "post": {
        "operationId": "AssessmentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssessmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "assessments"
        ]
      }
    },
    "/assessments/{id}": {
      "patch": {
        "operationId": "AssessmentsController_updateDraft",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "assessments"
        ]
      }
    },
...
    "/public/anamnese/{token}": {
      "get": {
        "operationId": "PublicIntakesController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "public-intakes"
        ]
      },
      "post": {
        "operationId": "PublicIntakesController_submit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitPublicIntakeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "public-intakes"
        ]
      }
    },
    "/public/intakes/invites": {
      "post": {
        "operationId": "PublicIntakesController_createInvite",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePublicInviteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "public-intakes"
        ]
      }
    },
    "/public/intakes/invites/{id}/revoke": {
      "post": {
        "operationId": "PublicIntakesController_revoke",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "public-intakes"
        ]
      }
    },
    "/public/intakes": {
      "get": {
        "operationId": "PublicIntakesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "public-intakes"
        ]
      }
    },
    "/public/intakes/{id}/approve": {
      "post": {
        "operationId": "PublicIntakesController_approve",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "public-intakes"
        ]
      }
    },
    "/public/intakes/{id}/merge": {
      "post": {
        "operationId": "PublicIntakesController_merge",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeIntakeRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "public-intakes"
        ]
      }
    },
    "/public/intakes/{id}/reject": {
      "post": {
        "operationId": "PublicIntakesController_reject",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectIntakeRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "public-intakes"
        ]
      }
    },
...
    "/replacement-links/{token}/reserve": {
      "post": {
        "operationId": "ReplacementLinksController_reserve",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReserveReplacementDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "replacement-links"
        ]
      }
    },
    "/replacement-links": {
      "post": {
        "operationId": "ReplacementLinksController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReplacementLinkDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "replacement-links"
        ]
      }
    },
    "/replacement-links/{id}/revoke": {
      "post": {
        "operationId": "ReplacementLinksController_revoke",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "replacement-links"
        ]
      }
    }
  },
  "info": {
    "title": "pilates-manager API",
    "description": "Backend API for Pilates studio management",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
...
      "CreateTemplateDto": {
        "type": "object",
        "properties": {}
      },
      "CreateAssessmentDto": {
        "type": "object",
        "properties": {}
      },
...
      "SubmitPublicIntakeDto": {
        "type": "object",
        "properties": {}
      },
      "CreatePublicInviteDto": {
        "type": "object",
        "properties": {}
      },
      "MergeIntakeRequestDto": {
        "type": "object",
        "properties": {}
      },
      "RejectIntakeRequestDto": {
        "type": "object",
        "properties": {}
      },
      "ReserveReplacementDto": {
        "type": "object",
        "properties": {}
      },
      "CreateReplacementLinkDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}